Should fix #54
Now everyone can create an account.
All characters from the provided api keys are stored in the db.
When updating user accesses, we look for positive standings on their characters (corps, alliances too). And put the users in the according groups.
Then, UrlPermission can do the rest of the job.